+++ src/pages/EditorPage/TabEditor/index.js
@@ -1,6 +1,6 @@
 import React, { useCallback, useEffect, useMemo, useState } from "react";
 import Editor from "@monaco-editor/react";
-import { Tab } from "./utils/const";
+import { Tab } from "../utils/const";
 
 export default function TabEditor({
   tab,
