+++ src/pages/EditorPage/Tabs/index.js
@@ -12,28 +12,15 @@ export default function Tabs({
   code,
 }) {
   return (
-    <ul
-      className={`nav nav-tabs`}
-      style={{
-        borderBottom: "0px",
-        marginTop: "9px",
-      }}
-    >
-      {isModule && (
-        <li className="nav-item">
-          <button
-            className={`nav-link ${
-              tab === Tab.Editor ? "active" : "text-secondary"
-            }`}
-            aria-current="page"
-            onClick={() => setTab(Tab.Editor)}
-          >
-            Module
-          </button>
-        </li>
-      )}
-      {isModule || (
-        <>
+    <div style={{ display: "flex" }}>
+      <ul
+        className={`nav nav-tabs`}
+        style={{
+          borderBottom: "0px",
+          marginTop: "9px",
+        }}
+      >
+        {isModule && (
           <li className="nav-item">
             <button
               className={`nav-link ${
@@ -42,52 +29,67 @@ export default function Tabs({
               aria-current="page"
               onClick={() => setTab(Tab.Editor)}
             >
-              Component
+              Module
             </button>
           </li>
-          <li className="nav-item">
-            <button
-              className={`nav-link ${
-                tab === Tab.Props ? "active" : "text-secondary"
-              }`}
-              aria-current="page"
-              onClick={() => setTab(Tab.Props)}
-            >
-              Props
-            </button>
-          </li>
-          {widgets.widgetMetadataEditor && (
+        )}
+        {isModule || (
+          <>
             <li className="nav-item">
               <button
                 className={`nav-link ${
-                  tab === Tab.Metadata ? "active" : "text-secondary"
+                  tab === Tab.Editor ? "active" : "text-secondary"
                 }`}
                 aria-current="page"
-                onClick={() => setTab(Tab.Metadata)}
+                onClick={() => setTab(Tab.Editor)}
               >
-                Metadata
+                Component
               </button>
             </li>
-          )}
-          {/* remove, and add show preview button on the right column */}
-          {layout === Layout.Tabs && (
             <li className="nav-item">
               <button
                 className={`nav-link ${
-                  tab === Tab.Widget ? "active" : "text-secondary"
+                  tab === Tab.Props ? "active" : "text-secondary"
                 }`}
                 aria-current="page"
-                onClick={() => {
-                  setRenderCode(code);
-                  setTab(Tab.Widget);
-                }}
+                onClick={() => setTab(Tab.Props)}
               >
-                Component Preview
+                Props
               </button>
             </li>
-          )}
-        </>
-      )}
-    </ul>
+            {widgets.widgetMetadataEditor && (
+              <li className="nav-item">
+                <button
+                  className={`nav-link ${
+                    tab === Tab.Metadata ? "active" : "text-secondary"
+                  }`}
+                  aria-current="page"
+                  onClick={() => setTab(Tab.Metadata)}
+                >
+                  Metadata
+                </button>
+              </li>
+            )}
+            {/* remove, and add show preview button on the right column */}
+            {layout === Layout.Tabs && (
+              <li className="nav-item">
+                <button
+                  className={`nav-link ${
+                    tab === Tab.Widget ? "active" : "text-secondary"
+                  }`}
+                  aria-current="page"
+                  onClick={() => {
+                    setRenderCode(code);
+                    setTab(Tab.Widget);
+                  }}
+                >
+                  Component Preview
+                </button>
+              </li>
+            )}
+          </>
+        )}
+      </ul>
+    </div>
   );
 }
